requirements.txt
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
